requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
spacy
scikit-learn
sentence-transformers
python-multipart
pdfminer.six
python-docx
numpy
pandas
gradio
sentence-transformers
pdfminer.six
matplotlib
pandas
reportlab